requirements.txt
torch
torchaudio
torchcodec
librosa
numpy
soundfile
scipy
pydub
pedalboard
onnxruntime
audio-separator
demucs
melband-roformer-infer
ml-collections
pyyaml
crepe
rvc
boto3
tqdm
